+++ NEWS.md
@@ -32,6 +32,12 @@
 
 * `tableby()`: Issue a warning when `coin` isn't available for the trend test. (#193)
 
+* `summary.freqlist()`: Implemented `head()` and `tail()`. (#188)
+
+* `summary.freqlist()`: fixed a bug when all table counts are 0 and `sparse=FALSE`. (#186, #194)
+
+* `freqlist()`: implemented a `sort()` method to sort tables on frequency. (#187)
+
 # arsenal v2.0.0
 
 There is a new class system (`"arsenal_table"`) which unifies `tableby()`, `modelsum()`, and `freqlist()`.